requirements.txt
aioredis==1.1.0
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.8.1
Automat==0.7.0
certifi==2018.8.13
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.2
Django==2.1.3
django-ipfs-storage==0.0.4
django-rest-swagger==2.2.0
djangorestframework==3.8.2
hiredis==0.2.0
humanize==0.5.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipfsapi==0.4.3
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
msgpack==0.5.6
netifaces==0.10.7
openapi-codec==1.3.2
pkg-resources==0.0.0
PyHamcrest==1.9.0
pytz==2018.5
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
six==1.11.0
Twisted==18.7.0
txaio==18.7.1
uritemplate==3.0.0
urllib3==1.23
zope.interface==4.5.0
